Fix spinner compilation error on visual studio

+++ external/gfx/imgui/imgui_spinner.cpp
@@ -29,7 +29,7 @@ namespace ImGui {
         window->DrawList->AddRectFilled(bb.Min, ImVec2(pos.x + circleStart, bb.Max.y), bg_col);
         window->DrawList->AddRectFilled(bb.Min, ImVec2(pos.x + circleStart*value, bb.Max.y), fg_col);
         
-        const float t = g.Time;
+        const float t = float(g.Time);
         const float r = size.y / 2;
         const float speed = 1.5f;
         
@@ -68,7 +68,7 @@ namespace ImGui {
         window->DrawList->PathClear();
         
         int num_segments = 30;
-        int start = abs(ImSin(g.Time*1.8f)*(num_segments-5));
+        int start = int(abs(float(ImSin(float(g.Time*1.8f))*(num_segments-5))));
         
         const float a_min = IM_PI*2.0f * ((float)start) / (float)num_segments;
         const float a_max = IM_PI*2.0f * ((float)num_segments-3) / (float)num_segments;
@@ -77,11 +77,11 @@ namespace ImGui {
         
         for (int i = 0; i < num_segments; i++) {
             const float a = a_min + ((float)i / (float)num_segments) * (a_max - a_min);
-            window->DrawList->PathLineTo(ImVec2(centre.x + ImCos(a+g.Time*8) * radius,
-                                                centre.y + ImSin(a+g.Time*8) * radius));
+            window->DrawList->PathLineTo(ImVec2(centre.x + ImCos(a+float(g.Time*8)) * radius,
+                                                centre.y + ImSin(a+float(g.Time*8)) * radius));
         }
 
-        window->DrawList->PathStroke(color, false, thickness);
+        window->DrawList->PathStroke(color, false, float(thickness));
         return true;
     }
 }
